+++ examples/basic/src/deps_full.cpp
@@ -1,4 +1,5 @@
 #include <AsyncUDP.h>
+#include <ArduinoOTA.h>
 #include <DNSServer.h>
 #include <ESPmDNS.h>
 #include <LittleFS.h>