+++ src/web/Esp32BaseWeb.h
@@ -62,6 +62,7 @@ public:
 
     static void setDefaultAuth(const char* user, const char* pass);
     static const char* authUser();
+    static const char* authPassword();
     static bool isAuthEnabled();
     static void setAuthEnabled(bool enabled);
     static bool checkAuth();